osmux: Improve code handling lost packets
With previous implementation, only up to a maximum of 6-7 rtp lost packets
can be replayed in best-case scenario. In other scenarios, big gaps can
easily occur, for instance if a lot of packets are lost when the current
batch is still empty.

If we forget to replay some lost RTP packets, then those are not taken into
account in the Osmux stream, which means sequence numbers and timestamps will
become different when Osmux re-generates the frames into RTP packets on the
other side, and some receivers may not play nice with this jumps in time,
eg. audio delays.

This new implementation tries to maintain status on the whole circuit
level rather than only on current batch. To simplify things, it clones
payload from latest packet to enqueue rather than last previosuly packet
enqueued.

Change-Id: I3b77d372cedadfd5588384bea0e5d3fd475195ce